+++ packages/formx/lib/src/extensions/form_field_state_extension.dart
@@ -4,6 +4,7 @@ import 'package:cross_file/cross_file.dart';
 import 'package:flutter/material.dart';
 
 import '../../formx.dart';
+import '../fields/time_formx_field.dart';
 
 /// Extends [FormFieldState] with a programatic way to set [errorText].
 extension FormFieldStateExtension<T> on FormFieldState<T> {
@@ -28,6 +29,10 @@ extension FormFieldStateExtension<T> on FormFieldState<T> {
       DateTime.tryParse(text) ??
       Localizations.of<MaterialLocalizations>(context, MaterialLocalizations)
           ?.parseCompactDate(text);
+
+  /// Returns the [FormFieldState.value] as a [DateTime].
+  TimeOfDay? get time =>
+      value?.tryCast<TimeOfDay>() ?? TimeFormxField.parseCompactTime(text);
 
   /// Whether [value] is empty.
   bool? get isEmpty {